@@ -5,6 +5,7 @@
   "author": "Benedikt Deicke <benedikt@userlist.com>",
   "homepage": "https://github.com/userlist/userlist-javascript",
   "license": "MIT",
+  "type": "module",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "files": [
@@ -13,6 +14,12 @@
     "src",
     "test"
   ],
+  "exports": {
+    ".": {
+      "import": "./dist/index.esm.js",
+      "require": "./dist/index.cjs.js"
+    }
+  },
   "publishConfig": {
     "access": "public"
   },
@@ -24,7 +31,7 @@
     "lint": "eslint src test",
     "build": "rollup -c",
     "serve": "rollup -c rollup.serve.js",
-    "test": "mocha --require ignore-styles --require test/helper --require @babel/register --recursive --colors",
+    "test": "rollup -c rollup.test.js",
     "test:watch": "pnpm test --watch",
     "prepublish": "pnpm lint && pnpm test && pnpm build"
   },
@@ -33,24 +40,36 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@babel/eslint-parser": "^7.26.10",
     "@babel/plugin-transform-runtime": "^7.9.0",
     "@babel/preset-env": "^7.9.0",
     "@babel/register": "^7.9.0",
     "@babel/runtime": "^7.9.2",
-    "@rollup/plugin-commonjs": "^11.0.2",
-    "@rollup/plugin-node-resolve": "^7.1.1",
-    "autoprefixer": "^9.7.5",
-    "chai": "^4.2.0",
-    "eslint": "6.8.0",
+    "@eslint/js": "^9.22.0",
+    "@rollup/plugin-babel": "^6.0.0",
+    "@rollup/plugin-commonjs": "^28.0.0",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "autoprefixer": "^10.0.0",
+    "chai": "^5.0.0",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "globals": "^16.0.0",
     "ignore-styles": "^5.0.1",
-    "jsdom": "^16.2.1",
-    "mocha": "^7.1.1",
+    "jsdom": "^26.0.0",
+    "mocha": "^11.0.0",
+    "postcss": "^8.5.3",
     "postcss-initial": "^3.0.2",
-    "rollup": "^2.2.0",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-postcss": "^2.5.0",
-    "rollup-plugin-serve": "^1.0.1",
-    "sinon": "^9.0.1",
-    "sinon-chai": "^3.5.0"
+    "prettier": "^3.5.3",
+    "rollup": "^4.0.0",
+    "rollup-plugin-import-css": "^3.5.8",
+    "rollup-plugin-mocha": "^0.0.1",
+    "rollup-plugin-postcss": "^4.0.0",
+    "rollup-plugin-serve": "^3.0.0",
+    "sinon": "^19.0.0",
+    "sinon-chai": "^4.0.0"
+  },
+  "dependencies": {
+    "@rollup/plugin-multi-entry": "^6.0.1"
   }
 }
