@@ -1,6 +1,6 @@
-var WorkerThreads = require('worker_threads');
+import { MessageChannel, MessagePort } from 'worker_threads';
 
-if(global.MessageChannel === undefined) {
-  global.MessageChannel = WorkerThreads.MessageChannel;
-  global.MessagePort = WorkerThreads.MessagePort;
+if (global.MessageChannel === undefined) {
+  global.MessageChannel = MessageChannel;
+  global.MessagePort = MessagePort;
 }
