@@ -1,39 +1,35 @@
-import babel from 'rollup-plugin-babel';
-import postcss from 'rollup-plugin-postcss';
-
+import babel from '@rollup/plugin-babel';
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
+import postcss from 'rollup-plugin-postcss';
 
 import autoprefixer from 'autoprefixer';
 import initial from 'postcss-initial';
 
-import pkg from './package.json';
+import pkg from './package.json' with { type: 'json' };
 
 export default {
   input: 'src/index.js',
   output: [
     {
       file: pkg.main,
-      format: 'cjs'
+      format: 'cjs',
     },
     {
       file: pkg.module,
-      format: 'esm'
-    }
+      format: 'esm',
+    },
   ],
   plugins: [
     resolve(),
     babel({
-      runtimeHelpers: true
+      babelHelpers: 'runtime',
     }),
     commonjs(),
     postcss({
       inject: false,
       minimize: true,
-      plugins: [
-        autoprefixer,
-        initial
-      ]
-    })
-  ]
+      plugins: [autoprefixer, initial],
+    }),
+  ],
 };
